feat(test_support): redefine recycle/1 to keep x-api-key
After a connection has been through a first request in the tests, it
needs to be recycled before to do a second request. Recycling erases
most things from the request, except a few request headers: `accept`,
`accept-language` and `authorization`.

By default, if we don’t do it ourselves in our tests,
`Phoenix.ConnTest.recycle/1` is called automatically on the connection
before making a new request. This is why it was previously working.

However, we are going to add a new required header: `x-api-key`. With
the default behaviour, it would be erase by `recycle/1`. Thankfully,
`recycle/2` allows us to define our own list of headers to keep. To
avoid repeating ourselves everywhere, let’s re-define `recycle/1` with
our own list, and explicitely use it in our tests. This will ensure
adding the requirement for `x-api-key` does not break our tests.

Refs: #1