test(c:profiles): make tests more stable
Tests where unstable because we matched against the presence of a number
in the HTML view to check whether a given ownership is present. The
issue is that the number representing the ID of the ownership could be
present for another reason in the page.

Adding some context to the match avoids matching any random number.

Refs: #50