chore(repo): rename all uuids to ids while keeping uuid type
The only important modification occurs in the migration file:
priv/repo/migrations/20240423224527_rename_all_uuid_to_id.exs

All the rest is mass renaming to adapt code to database fields names.

We could later remove some options we needed when primary key was
named 'uuid' instead of 'id'.
I think of:
* returning: [:id] (option passed when inserting new object to the database)
* @derive {Phoenix.Param, key: :id} (option passed in schemas)
* param: "id" (option passed in controllers)

As these options brang more clarity to the code, I chose to don’t remove
them for now. It could be done in a distinct commit, to have a clear history
vision if we chose to remove them.

Refs: #18