refactor!: move all contexts under `Prepair.LegacyContexts`
This is the first commit of the migration to the Ash framework. In order
to prepair for the migration of the Phoenix contexts to Ash domains, we
first move all contexts under a LegacyContexts submodule. This way,
we’ll be able to add similar Ash domains in parallel, and avoid breaking
existing code and tests during the migration process.

Refs: #80

BREAKING CHANGE: All contexts are now under `Prepair.LegacyContexts`.